requirements.txt
discord-py-interactions
discord-py
flask
